+++ src/index.js
@@ -4,6 +4,7 @@ export { default as useWillUnmount } from './useWillUnmount';
 export { default as useLifecycle } from './useLifecycle';
 export { default as useWindowResize } from './useWindowResize';
 export { default as useDebouncedCallback } from './useDebouncedCallback';
+export { default as useThrottledCallback } from './useThrottledCallback';
 export { default as useMouse } from './useMouse';
 export { default as useMouseHandler } from './useMouseHandler';
 export { default as useMouseState } from './useMouseState';